// src/app/models/academics/student-application-category-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * StudentApplicationCategoryView
 */
export interface StudentApplicationCategoryView { 
    /**
     * Gets or sets
     */
    name: string;
    id: string;
}